+++ ThermostatApp/app/src/main/java/a2id40/thermostatapp/activities/base/BaseActivity.java
@@ -12,10 +12,11 @@ import android.view.MenuItem;
 
 import a2id40.thermostatapp.R;
 import a2id40.thermostatapp.activities.base.util.ActivityUtils;
-import a2id40.thermostatapp.fragments.daynight.DayNightFragment;
 import a2id40.thermostatapp.fragments.help.HelpFragment;
 import a2id40.thermostatapp.fragments.main.MainFragment;
+import a2id40.thermostatapp.fragments.settings.SettingsFragment;
 import a2id40.thermostatapp.fragments.vacation.VacationFragment;
+import a2id40.thermostatapp.fragments.viewweekly.ViewWeeklyFragment;
 import a2id40.thermostatapp.fragments.weekly.WeeklyFragment;
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -83,8 +84,8 @@ public class BaseActivity extends AppCompatActivity implements NavigationView.On
             case R.id.base_drawer_main:
                 ActivityUtils.replaceFragment(getSupportFragmentManager(), MainFragment.newInstance(), R.id.activity_base_container);
                 break;
-            case R.id.base_drawer_daynight:
-                ActivityUtils.replaceFragment(getSupportFragmentManager(), DayNightFragment.newInstance(), R.id.activity_base_container);
+            case R.id.base_drawer_viewWeekly:
+                ActivityUtils.replaceFragment(getSupportFragmentManager(), ViewWeeklyFragment.newInstance(), R.id.activity_base_container);
                 break;
             case R.id.base_drawer_help:
                 ActivityUtils.replaceFragment(getSupportFragmentManager(), HelpFragment.newInstance(), R.id.activity_base_container);
@@ -94,6 +95,9 @@ public class BaseActivity extends AppCompatActivity implements NavigationView.On
                 break;
             case R.id.base_drawer_weekly:
                 ActivityUtils.replaceFragment(getSupportFragmentManager(), WeeklyFragment.newInstance(), R.id.activity_base_container);
+                break;
+            case R.id.base_drawer_settings:
+                ActivityUtils.replaceFragment(getSupportFragmentManager(), SettingsFragment.newInstance(), R.id.activity_base_container);
                 break;
             default:
                 break;
