+++ ThermostatApp/app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="DAY NIGHT FRAGMENT"
+        android:text="SETTINGS FRAGMENT"
         android:textSize="30dp"
         android:gravity="center"/>
 </LinearLayout>
