+++ ThermostatApp/app/src/main/res/layout/fragment_viewweekly.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="VIEW WEEKLY FRAGMENT"
+        android:textSize="30dp"
+        android:gravity="center"/>
+</LinearLayout>
