+++ ThermostatApp/app/src/main/res/menu/activity_base_drawer.xml
@@ -11,13 +11,17 @@
             android:icon="@drawable/ic_menu_manage"
             android:title="@string/activity_base_drawer_weekly" />
         <item
-            android:id="@+id/base_drawer_daynight"
+            android:id="@+id/base_drawer_viewWeekly"
             android:icon="@drawable/ic_menu_manage"
-            android:title="@string/activity_base_drawer_daynight" />
+            android:title="@string/activity_base_drawer_viewWeekly" />
         <item
             android:id="@+id/base_drawer_vacation"
             android:icon="@drawable/ic_menu_manage"
             android:title="@string/activity_base_drawer_vaction" />
+        <item
+            android:id="@+id/base_drawer_settings"
+            android:icon="@drawable/ic_menu_manage"
+            android:title="@string/activity_base_drawer_settings" />
         <item
             android:id="@+id/base_drawer_help"
             android:icon="@drawable/ic_menu_manage"
