+++ ThermostatApp/app/src/main/res/values/strings.xml
@@ -5,9 +5,10 @@
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="action_settings">Settings</string>
-    <string name="activity_base_drawer_weekly">Weekly Program</string>
-    <string name="activity_base_drawer_daynight">Day and Night Settings</string>
-    <string name="activity_base_drawer_vaction">On Vacation Program</string>
+    <string name="activity_base_drawer_weekly">Set Weekly</string>
+    <string name="activity_base_drawer_viewWeekly">View Weekly</string>
+    <string name="activity_base_drawer_vaction">Set Vacation</string>
+    <string name="activity_base_drawer_settings">Settings</string>
     <string name="activity_base_drawer_help">Help</string>
     <string name="activity_base_drawer_header_title">Thermostat App</string>
     <string name="activity_base_drawer_main">Main</string>
